#334: Extract date helpers into helper file
